Monitor: fix a few warnings
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@45043 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/monitor/mon_command.c
@@ -973,7 +973,7 @@ static void print_help_topics(const char *filter)
     int len;
     int longest;
     int max_col;
-    int filter_len = 0;
+    size_t filter_len = 0;
     int isfirst = 1;
     int count = 0;
     int cmd_num = 0;
